Add experimental buildah support
A new optional flag called --manager can be set to buildah to try out
using buildah for building images and exporting the results. This is
currently experimental and shouldn't be used for real image construction
yet.

By default the flag is set to moby and will use moby/docker as it's
image manager.